move the source to the core
