+++ martin-core/src/tiles/postgres/errors.rs
@@ -1,15 +1,14 @@
 //! Error types for `PostgreSQL` operations.
 
-use std::collections::HashMap;
 use std::io;
 use std::path::PathBuf;
 
 use deadpool_postgres::tokio_postgres::Error as TokioPgError;
 use deadpool_postgres::{BuildError, PoolError};
 use martin_tile_utils::TileCoord;
-use postgres::types::Json;
 use semver::Version;
 
+use super::utils::query_to_json;
 use crate::tiles::UrlQuery;
 
 /// Result type for `PostgreSQL` operations.
@@ -97,19 +96,4 @@ pub enum PgError {
     /// Configuration error.
     #[error("Configuration error: {0}")]
     ConfigError(&'static str),
-}
-
-#[must_use]
-pub fn query_to_json(query: Option<&UrlQuery>) -> Json<HashMap<String, serde_json::Value>> {
-    let mut query_as_json = HashMap::new();
-    if let Some(query) = query {
-        for (k, v) in query {
-            let json_value: serde_json::Value =
-                serde_json::from_str(v).unwrap_or_else(|_| serde_json::Value::String(v.clone()));
-
-            query_as_json.insert(k.clone(), json_value);
-        }
-    }
-
-    Json(query_as_json)
 }
