+++ martin-core/src/tiles/postgres/mod.rs
@@ -5,3 +5,8 @@ mod tls;
 
 mod pool;
 pub use pool::PgPool;
+
+mod source;
+pub use source::{PgSource, PgSqlInfo};
+
+pub(crate) mod utils;
