+++ martin/src/pg/builder.rs
@@ -7,7 +7,7 @@ use log::{debug, error, info, warn};
 use martin_core::config::OptBoolObj::{Bool, NoValue, Object};
 use martin_core::config::OptOneMany::NoVals;
 use martin_core::tiles::BoxedSource;
-use martin_core::tiles::postgres::{PgError, PgPool, PgResult};
+use martin_core::tiles::postgres::{PgError, PgPool, PgResult, PgSource, PgSqlInfo};
 
 use crate::config::args::BoundsCalcType;
 use crate::config::file::pg::{
@@ -16,7 +16,6 @@ use crate::config::file::pg::{
 };
 use crate::pg::query_functions::query_available_function;
 use crate::pg::query_tables::{query_available_tables, table_to_query};
-use crate::pg::source::{PgSource, PgSqlInfo};
 use crate::pg::utils::{find_info, find_kv_ignore_case, normalize_key};
 use crate::utils::IdResolver;
