+++ martin/src/pg/mod.rs
@@ -1,7 +1,6 @@
 pub(crate) mod builder;
 mod query_functions;
 mod query_tables;
-mod source;
 pub(crate) mod utils;
 
 pub use query_functions::query_available_function;
