+++ martin/src/pg/query_functions.rs
@@ -5,13 +5,12 @@ use std::iter::zip;
 
 use log::{debug, warn};
 use martin_core::tiles::postgres::PgError::PostgresError;
-use martin_core::tiles::postgres::{PgPool, PgResult};
+use martin_core::tiles::postgres::{PgPool, PgResult, PgSqlInfo};
 use postgres_protocol::escape::escape_identifier;
 use serde_json::Value;
 
 use crate::config::file::pg::FunctionInfo;
 use crate::pg::builder::SqlFuncInfoMapMap;
-use crate::pg::source::PgSqlInfo;
 
 /// Queries the database for available tile-generating functions.
 ///
