+++ martin/src/pg/query_tables.rs
@@ -5,7 +5,7 @@ use std::collections::HashMap;
 use futures::pin_mut;
 use log::{debug, warn};
 use martin_core::tiles::postgres::PgError::PostgresError;
-use martin_core::tiles::postgres::{PgPool, PgResult};
+use martin_core::tiles::postgres::{PgPool, PgResult, PgSqlInfo};
 use martin_tile_utils::EARTH_CIRCUMFERENCE_DEGREES;
 use postgis::ewkb;
 use postgres_protocol::escape::{escape_identifier, escape_literal};
@@ -16,7 +16,6 @@ use tokio::time::timeout;
 use crate::config::args::{BoundsCalcType, DEFAULT_BOUNDS_TIMEOUT};
 use crate::config::file::pg::{PgInfo, TableInfo};
 use crate::pg::builder::SqlTableInfoMapMapMap;
-use crate::pg::source::PgSqlInfo;
 
 static DEFAULT_EXTENT: u32 = 4096;
 static DEFAULT_BUFFER: u32 = 64;
